+++ lib/core/const/const.dart
@@ -6,6 +6,8 @@ const String quizScreen="/quiz-screen";
 const String userData="/user-data";
 const String chatScreen="/chat-screen";
 const String studentRegisterScreen="/studentRegister-screen";
+const String homeDeciderScreen="/homeDecider-screen";
+
