+++ lib/di/di.dart
@@ -21,9 +21,9 @@ import '../features/quiz/data/rebo/add_impl.dart';
 import '../features/register_screen/cubit/register_cubit.dart';
 import '../features/register_screen/data/rebo/rebo.dart';
 import '../features/register_screen/data/rebo/rebo_impl.dart';
-import '../features/student/data/rebo/register_repo.dart';
-import '../features/student/data/rebo/register_repo_impl.dart';
-import '../features/student/logic/student_cubit.dart';
+import '../features/type_register/data/rebo/register_repo.dart';
+import '../features/type_register/data/rebo/register_repo_impl.dart';
+import '../features/type_register/logic/type_register_cubit.dart';
 
 final sl = GetIt.instance;
 
@@ -50,6 +50,6 @@ void setupServiceLocator() {
   sl.registerFactory<HomeCubit>(() => HomeCubit(sl<GetAllQues>()));
   sl.registerFactory<SaveScoreCubit>(() => SaveScoreCubit(sl<AddScore>(), sl<UpdateUserDataCubit>()));
   sl.registerLazySingleton<UpdateUserDataCubit>(() => UpdateUserDataCubit(sl<UpdateUserRebo>()));
-  sl.registerLazySingleton<StudentRegisterCubit>(() => StudentRegisterCubit(sl<RegisterRepository>(),sl<FirebaseAuth>()));
+  sl.registerFactory<TypeRegisterCubit>(() => TypeRegisterCubit(sl<RegisterRepository>(),sl<FirebaseAuth>()));
 
 }
