+++ lib/features/login_screen/logic/cubit/login_cubit.dart
@@ -43,6 +43,12 @@ class LoginCubit extends Cubit<LoginStates> {
           )
           .then((value) async {
             await CacheHelper.putString(key: "uid", value: value.user!.uid);
+            await FirebaseFirestore.instance
+                .collection('users')
+                .doc(value.user!.uid)
+                .get().then((value) {
+                  CacheHelper.putString(key: "type", value: value.data()!["type"]);
+            });
             emit(AuthSuccess());
 
       });
@@ -58,6 +64,8 @@ class LoginCubit extends Cubit<LoginStates> {
         .signOut()
         .then((onValue) {
           CacheHelper.removeString(key: "uid");
+          CacheHelper.removeString(key: "type");
+          CacheHelper.removeString(key: "submitted");
           emit(SignOutDone());
         })
         .catchError((onError) {
