+++ lib/features/student/presentation/student_home.dart
@@ -1,0 +1,17 @@
+import 'package:add_ques/core/theme/colors/colors.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class StudentHomeScreen extends StatelessWidget {
+  const StudentHomeScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: AppColors.mainTealColor,
+      body: Center(
+        child: Text("Student Home"),
+      ),
+    );
+  }
+}
