+++ lib/features/type_register/data/models/register_model.dart
@@ -1,0 +1,27 @@
+abstract class RegisterModel {
+  final String fullName;
+  final String phone;
+  final String gender;
+  final String status;
+  final String dateOfBirth;
+  final String city;
+  final String imageUrl;
+  final String university;
+  final String email;
+
+  RegisterModel({
+    required this.fullName,
+    required this.university,
+    required this.status,
+    required this.phone,
+    required this.gender,
+    required this.dateOfBirth,
+    required this.city,
+    required this.email,
+    required this.imageUrl,
+  });
+
+}
+
+
+
