+++ lib/features/type_register/data/rebo/register_repo.dart
@@ -1,3 +1,5 @@
+import '../../../doctor/data/models/doctor_model.dart';
+import '../../../student/daya/models/student_model.dart';
 import '../models/register_model.dart';
 
 abstract class RegisterRepository {
