+++ lib/features/type_register/logic/type_register_state.dart
@@ -1,0 +1,26 @@
+
+abstract class TypeRegisterState {}
+
+class StudentInitial extends TypeRegisterState {}
+
+class NextStep extends TypeRegisterState {}
+
+class PreviousStep extends TypeRegisterState {}
+
+class SetCurrent extends TypeRegisterState {}
+
+class StudentGenderChanged extends TypeRegisterState{}
+
+class PickDateDone extends TypeRegisterState{}
+
+class StudentSubmitted extends TypeRegisterState{}
+
+
+class StudentSubmitLoading extends TypeRegisterState{}
+
+class StudentSubmitFailed extends TypeRegisterState{
+  String error;
+  StudentSubmitFailed({required this.error});
+}
+
+
