+++ lib/features/type_register/presentation/type_register_ui.dart
@@ -3,27 +3,27 @@ import 'package:add_ques/core/shared_widgets/shared_text_form_field.dart';
 import 'package:add_ques/core/theme/colors/colors.dart';
 import 'package:add_ques/core/theme/text_themes/text.dart';
 import 'package:add_ques/features/register_screen/cubit/register_cubit.dart';
-import 'package:add_ques/features/student/presentation/widgets/step1_name_phone.dart';
-import 'package:add_ques/features/student/presentation/widgets/step2_year_city_uni.dart';
-import 'package:add_ques/features/student/presentation/widgets/step3_doctor.dart';
-import 'package:add_ques/features/student/presentation/widgets/step3_student.dart';
+import 'package:add_ques/features/type_register/presentation/widgets/step1_name_phone.dart';
+import 'package:add_ques/features/type_register/presentation/widgets/step2_year_city_uni.dart';
+import 'package:add_ques/features/type_register/presentation/widgets/step3_doctor.dart';
+import 'package:add_ques/features/type_register/presentation/widgets/step3_student.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 import '../../../core/const/const.dart';
 import '../../../core/shared_widgets/snack_bar.dart';
-import '../logic/student_cubit.dart';
-import '../logic/student_state.dart';
+import '../logic/type_register_cubit.dart';
+import '../logic/type_register_state.dart';
 
-class StepperExample extends StatelessWidget {
+class TypeRegisterScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return SafeArea(
       child: Scaffold(
-        body: BlocBuilder<StudentRegisterCubit, StudentState>(
+        body: BlocBuilder<TypeRegisterCubit, TypeRegisterState>(
           builder: (context, state) {
-            var cubit = BlocProvider.of<StudentRegisterCubit>(context);
+            var cubit = BlocProvider.of<TypeRegisterCubit>(context);
             return Theme(
               data: Theme.of(context).copyWith(
                 colorScheme: ColorScheme.light(
@@ -111,7 +111,7 @@ class StepperExample extends StatelessWidget {
                       ],
                     ),
                   ),
-                  BlocListener<StudentRegisterCubit, StudentState>(
+                  BlocListener<TypeRegisterCubit, TypeRegisterState>(
                     listener: (BuildContext context, state) {
                       if (state is StudentSubmitLoading) {
                         showDialog(
@@ -126,7 +126,7 @@ class StepperExample extends StatelessWidget {
                         );
                       }
                       if (state is StudentSubmitted) {
-                        context.pushAndRemoveUntil(homeScreen);
+                        context.pushAndRemoveUntil(homeDeciderScreen);
                       }
                       if (state is StudentSubmitFailed) {
                         AwesomeSnackBar(
