+++ lib/features/type_register/presentation/widgets/step3_doctor.dart
@@ -6,14 +6,14 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import '../../../../core/shared_widgets/shared_text_form_field.dart';
 import '../../../../core/theme/colors/colors.dart';
 import '../../../../core/theme/text_themes/text.dart';
-import '../../logic/student_cubit.dart';
+import '../../logic/type_register_cubit.dart';
 
-class SetStep3ForStudent extends StatelessWidget {
-  const SetStep3ForStudent({super.key});
+class SetStep3ForDoctors extends StatelessWidget {
+  const SetStep3ForDoctors({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final cubit = BlocProvider.of<StudentRegisterCubit>(context);
+    final cubit = BlocProvider.of<TypeRegisterCubit>(context);
     return Padding(
       padding: EdgeInsets.all(20.0.sign),
       child: Form(
@@ -23,58 +23,63 @@ class SetStep3ForStudent extends StatelessWidget {
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
             Text(
-              "Enter Your grade or year of graduate",
+              "Enter Your Years of experience",
               style: TextThemes.font16BlackBold.copyWith(
                 color: AppColors.mainTealColor,
               ),
             ),
             SizedBox(height: 15.h),
             SharedTextFormField(
-              hintText: "Enter Your grade or year of graduate",
+              hintText: "Enter Your Years of experience",
               validator: (validator){
                 if(validator!.isEmpty) {
-                  return  "Enter Your grade or year of graduate";
+                  return  "Enter Your Years of experience";
                 }
                 return null;
               },
-              controller: cubit.gradeController,
+              controller: cubit.expController,
             ),
             SizedBox(height: 20.h,),
             Text(
-              "Enter Your Department  ",
+              "Enter Your Specialization ",
               style: TextThemes.font16BlackBold.copyWith(
                 color: AppColors.mainTealColor,
               ),
             ),
             SizedBox(height: 15.h),
             SharedTextFormField(
-              hintText: "Enter Your Department ",
+              hintText: "Enter Your Specialization",
               validator: (validator) {
                 if(validator!.isEmpty) {
-                  return "Enter Your Department ";
+                  return "Enter Your Specialization";
                 }
                 return null;
               },
-              controller: cubit.departmentController ,
+              controller: cubit.specializationController,
             ),
             SizedBox(height: 20.h,),
             Text(
-              "Enter your Gpa",
+              "Enter your Graduation year",
               style: TextThemes.font16BlackBold.copyWith(
                 color: AppColors.mainTealColor,
               ),
             ),
             SizedBox(height: 15.h),
-            SharedTextFormField(
-              suffixIcon: Icon(Icons.calendar_today),
-              hintText: "Enter your Gpa",
-              validator: (validator) {
-                if(validator!.isEmpty) {
-                  return "Enter your Gpa";
-                }
-                return null;
+            GestureDetector(
+              onTap: (){
+                cubit.selectDate(context);
               },
-              controller: cubit.gpaController,
+              child: SharedTextFormField(
+                suffixIcon: Icon(Icons.calendar_today),
+                hintText: "Enter your Graduation year",
+                validator: (validator) {
+                  if(validator!.isEmpty) {
+                    return "Enter your Graduation year";
+                  }
+                  return null;
+                },
+                controller: cubit.graduationController,
+              ),
             ),
           ],
         ),
