+++ lib/features/type_register/presentation/widgets/step3_student.dart
@@ -6,14 +6,14 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import '../../../../core/shared_widgets/shared_text_form_field.dart';
 import '../../../../core/theme/colors/colors.dart';
 import '../../../../core/theme/text_themes/text.dart';
-import '../../logic/student_cubit.dart';
+import '../../logic/type_register_cubit.dart';
 
-class SetStep3ForDoctors extends StatelessWidget {
-  const SetStep3ForDoctors({super.key});
+class SetStep3ForStudent extends StatelessWidget {
+  const SetStep3ForStudent({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final cubit = BlocProvider.of<StudentRegisterCubit>(context);
+    final cubit = BlocProvider.of<TypeRegisterCubit>(context);
     return Padding(
       padding: EdgeInsets.all(20.0.sign),
       child: Form(
@@ -23,43 +23,43 @@ class SetStep3ForDoctors extends StatelessWidget {
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
             Text(
-              "Enter Your Years of experience",
+              "Enter Your grade or year of graduate",
               style: TextThemes.font16BlackBold.copyWith(
                 color: AppColors.mainTealColor,
               ),
             ),
             SizedBox(height: 15.h),
             SharedTextFormField(
-              hintText: "Enter Your Years of experience",
+              hintText: "Enter Your grade or year of graduate",
               validator: (validator){
                 if(validator!.isEmpty) {
-                  return  "Enter Your Years of experience";
+                  return  "Enter Your grade or year of graduate";
                 }
                 return null;
               },
-              controller: cubit.expController,
+              controller: cubit.gradeController,
             ),
             SizedBox(height: 20.h,),
             Text(
-              "Enter Your Specialization ",
+              "Enter Your Department  ",
               style: TextThemes.font16BlackBold.copyWith(
                 color: AppColors.mainTealColor,
               ),
             ),
             SizedBox(height: 15.h),
             SharedTextFormField(
-              hintText: "Enter Your Specialization",
+              hintText: "Enter Your Department ",
               validator: (validator) {
                 if(validator!.isEmpty) {
-                  return "Enter Your Specialization";
+                  return "Enter Your Department ";
                 }
                 return null;
               },
-              controller: cubit.specializationController,
+              controller: cubit.departmentController ,
             ),
             SizedBox(height: 20.h,),
             Text(
-              "Enter your Graduation year",
+              "Enter your Gpa",
               style: TextThemes.font16BlackBold.copyWith(
                 color: AppColors.mainTealColor,
               ),
@@ -67,14 +67,14 @@ class SetStep3ForDoctors extends StatelessWidget {
             SizedBox(height: 15.h),
             SharedTextFormField(
               suffixIcon: Icon(Icons.calendar_today),
-              hintText: "Enter your Graduation year",
+              hintText: "Enter your Gpa",
               validator: (validator) {
                 if(validator!.isEmpty) {
-                  return "Enter your Graduation year";
+                  return "Enter your Gpa";
                 }
                 return null;
               },
-              controller: cubit.graduationController,
+              controller: cubit.gpaController,
             ),
           ],
         ),
